ServerCore is now compiling
