+++ src/Compilers/Server/Core/Assembly.cs
