+++ src/Compilers/Server/Core/DiagnosticListener.cs
