+++ src/Compilers/Server/Core/ServerCore.csproj
@@ -37,15 +37,12 @@
     <ErrorReport>prompt</ErrorReport>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Assembly.cs" />
-    <Compile Include="..\..\Shared\ShadowCopyAnalyzerAssemblyLoader.cs">
-      <Link>ShadowCopyAnalyzerAssemblyLoader.cs</Link>
-    </Compile>
-    <Compile Include="..\..\Shared\BuildProtocol.cs">
-      <Link>BuildProtocol.cs</Link>
-    </Compile>
     <Compile Include="..\..\Shared\CompilerServerLogger.cs">
       <Link>CompilerServerLogger.cs</Link>
+    </Compile>
+    <Compile Include="Assembly.cs" />
+    <Compile Include="..\..\Shared\BuildProtocol.cs">
+      <Link>BuildProtocol.cs</Link>
     </Compile>
     <Compile Include="CompilerRequestHandler.cs" />
     <Compile Include="CSharpCompilerServer.cs" />
