+++ src/Compilers/Server/Core/ServerDispatcher.MemoryHelper.cs
