+++ src/Compilers/Server/Core/project.json
