--- src/Compilers/Server/ServerCore/ICompilerServerHost.cs
@@ -1,14 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Microsoft.CodeAnalysis.CompilerServer
-{
-    class ICompilerServerHost
-    {
-    }
-}
