+++ src/Compilers/Server/VBCSCompiler/VBCSCompiler.csproj
@@ -31,7 +31,7 @@
       <Project>{1EE8CAD3-55F9-4D91-96B2-084641DA9A6C}</Project>
       <Name>CodeAnalysis</Name>
     </ProjectReference>
-    <ProjectReference Include="..\ServerCore\ServerCore.csproj">
+    <ProjectReference Include="..\Core\ServerCore.csproj">
       <Project>{9929e5ae-85ca-46a8-a500-3be7bc634d87}</Project>
       <Name>ServerCore</Name>
     </ProjectReference>
@@ -55,8 +55,15 @@
     <Compile Include="..\..\Shared\AnalyzerAssemblyLoadUtils.cs">
       <Link>AnalyzerAssemblyLoadUtils.cs</Link>
     </Compile>
+    <Compile Include="..\..\Shared\CompilerServerLogger.cs">
+      <Link>CompilerServerLogger.cs</Link>
+    </Compile>
+    <Compile Include="..\..\Shared\ShadowCopyAnalyzerAssemblyLoader.cs">
+      <Link>ShadowCopyAnalyzerAssemblyLoader.cs</Link>
+    </Compile>
     <Compile Include="AnalyzerConsistencyChecker.cs" />
     <Compile Include="Assembly.cs" />
+    <Compile Include="DesktopCompilerServerHost.cs" />
     <Compile Include="NamedPipeClientConnection.cs" />
     <Compile Include="VBCSCompiler.cs" />
   </ItemGroup>
